VideoCore: Move Regs to its own file
