+++ src/video_core/clipper.cpp
@@ -12,10 +12,10 @@
 #include "common/logging/log.h"
 #include "common/vector_math.h"
 #include "video_core/clipper.h"
-#include "video_core/pica.h"
 #include "video_core/pica_state.h"
 #include "video_core/pica_types.h"
 #include "video_core/rasterizer.h"
+#include "video_core/regs.h"
 #include "video_core/shader/shader.h"
 
 using Pica::Rasterizer::Vertex;
