+++ src/video_core/regs_lighting.h
@@ -6,9 +6,11 @@
 
 #include <array>
 
+#include "common/assert.h"
 #include "common/bit_field.h"
 #include "common/common_funcs.h"
 #include "common/common_types.h"
+#include "common/vector_math.h"
 
 namespace Pica {
